+++ GHEtool/test/unit-tests/test_main_class.py
@@ -6,7 +6,7 @@ import numpy as np
 import pygfunction as gt
 import pytest
 
-from GHEtool import GroundConstantTemperature, GroundFluxTemperature, FluidData, PipeData, Borefield, SizingSetup, FOLDER
+from GHEtool import GroundConstantTemperature, GroundFluxTemperature, FluidData, PipeData, Borefield, CalculationSetup, FOLDER
 from GHEtool.logger import ghe_logger
 from GHEtool.Validation.cases import load_case
 from GHEtool.VariableClasses.LoadData import MonthlyGeothermalLoadAbsolute, HourlyGeothermalLoad
@@ -342,19 +342,14 @@ def test_Carcel(ground_data, constant_Rb, result):
 def test_set_sizing_setup():
     borefield = Borefield()
     sizing_setup_backup = copy.deepcopy(borefield._sizing_setup)
-    H_init_backup = borefield.H_init
     borefield.sizing_setup()
     assert borefield._sizing_setup == sizing_setup_backup
-    assert borefield.H_init == H_init_backup
     # set sizing_setup
-    test = SizingSetup(4, False, True, False)
-    test2 = SizingSetup(3, False, False, True)
-    borefield.sizing_setup(120, True, 4, False, True, False)
-    assert borefield.H_init == 120
-    assert borefield._sizing_setup == test
+    test = CalculationSetup(4, False, True, False)
+    test2 = CalculationSetup(3, False, False, True)
+    borefield.sizing_setup(use_constant_Rb=True, quadrant_sizing=4, L2_sizing=False, L3_sizing=True, L4_sizing=False)
     assert borefield._sizing_setup == test
     borefield.sizing_setup(sizing_setup=test2)
-    assert borefield.H_init == 120
     assert borefield._sizing_setup == test2
 
 
@@ -871,3 +866,26 @@ def test_optimise_load_profile_without_hourly_data():
     borefield.load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"))
     borefield.create_rectangular_borefield(10, 10, 6, 6, 150)
     borefield.optimise_load_profile(borefield.load)
+
+@pytest.mark.parametrize("H, result",
+                         zip(range(110, 130, 1), [122.99210426454648, 122.99135446500962, 122.99135409065917,
+                                                  122.99135403971272, 122.99135403719148, 122.9913540367823,
+                                                  122.99135403674013, 122.99135403673134, 122.99221686744185,
+                                                  122.99217229220649, 122.99135553171544, 122.99220727438545,
+                                                  122.99477143007601, 122.9921794642058, 122.99220615327106,
+                                                  122.99221262582992, 122.99221900364599, 122.9922252887964,
+                                                  122.99223148329897, 122.99223758911434]))
+def test_effect_H_init(H, result):
+    borefield = Borefield()
+    borefield.ground_data = GroundConstantTemperature(3, 11)
+    borefield.create_rectangular_borefield(10, 5, 7, 7, 100, 0.75)
+    load = MonthlyGeothermalLoadAbsolute(*load_case(1))
+    borefield.load = load
+    borefield.sizing_setup(H_init=H)
+    assert np.isclose(borefield.size(), result)
+    borefield = Borefield()
+    borefield.ground_data = GroundConstantTemperature(3, 11)
+    borefield.create_rectangular_borefield(10, 5, 7, 7, 100, 0.75)
+    load = MonthlyGeothermalLoadAbsolute(*load_case(1))
+    borefield.load = load
+    assert np.isclose(borefield.size(H_init=H), result)